1.增加copytrade、PSPmint、Stake、apiKey创建脚本；2.批量创建用户和登录获取token；3.并发脚本 4.添加压测依赖

+++ src/main/java/com/phemex/dataFactory/common/utils/BytesUtils.java
@@ -20,7 +20,7 @@ public class BytesUtils {
 
         for(int i = 7; i >= 0; --i) {
             res <<= 8;
-            res |= (long)(b[i] & 255);
+            res |= b[i] & 255;
         }
 
         return res;
